Fix e2e test + prepare new one
