+++ e2e-tests/src/test/java/eu/adainius/newsfocused/e2e/E2ETestPatternTester.java
@@ -24,23 +24,14 @@ import eu.adainius.newsfocused.user.User;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-public class E2ETest {
+public class E2ETestPatternTester {
     private static String buildMailcatcherDocker = "docker build --rm -t news_focused/mailcatcher ./src/test/resources/mailServer";
     private static String runMailcatcherDocker = "docker run --rm --name=news_focused-mailcatcher -d --publish=1080:1080 --publish=10025:10025 news_focused/mailcatcher";
     private static String stopMailcatcherDocker = "docker stop news_focused-mailcatcher";
 
     @BeforeAll
     public static void runMailServer() throws IOException, InterruptedException {
-        stopMailServer();
-        Process p;
-        Runtime rt = Runtime.getRuntime();
-        p = rt.exec(buildMailcatcherDocker);
-        p.waitFor();
-        logProcess(p);
-        p = Runtime.getRuntime().exec(runMailcatcherDocker);
-        p.waitFor();
-        logProcess(p);
-        log.info("Started mail server");
+
     }
 
     private static void logProcess(Process p) throws IOException {
@@ -50,40 +41,21 @@ public class E2ETest {
 
     @AfterAll
     public static void stopMailServer() throws IOException, InterruptedException {
-        log.info("Stopping mail server");
-        Process p = Runtime.getRuntime().exec(stopMailcatcherDocker);
-        p.waitFor();
-        logProcess(p);
+
     }
 
     @Test
     public void parses_headlines_from_BBC_and_sends_email(@TempDir Path tempDir) throws Exception {
-        HttpClient httpClient = HttpClient.newHttpClient();
-        String mailServerUrl = "http://127.0.0.1:1080/";
+        // GIVEN
 
-        String usersFile = "src/test/resources/users.json";
-        String repoLocation = createRepoLocation(tempDir);
 
-        App.main(new String[] { usersFile, repoLocation });
+        // WHEN
+        
 
-        assertEmailWasSent(httpClient, mailServerUrl);
-    }
 
-    private void assertEmailWasSent(HttpClient httpClient, String mailServerUrl)
-            throws IOException, InterruptedException {
-        var request = HttpRequest.newBuilder(URI.create(mailServerUrl + "messages"))
-                .header("accept", "application/json").build();
+        // THEN
 
-        // use the client to send the request
-        var response = httpClient.send(request, BodyHandlers.ofString());
 
-        JsonArray messageArray = new Gson().fromJson(response.body(), JsonArray.class);
-
-        assertTrue(messageArray.size() > 0);
-    }
-
-    private String createRepoLocation(Path tempDir) throws IOException {
-        return Files.createDirectories(tempDir.resolve("repo")).toString();
     }
 
 }
